requirements.txt
# Core RAG dependencies
langchain==0.1.0
langchain-community==0.0.10
langchain-openai==0.0.5
chromadb==0.4.22
sentence-transformers==2.2.2

# LLM frameworks
transformers==4.36.2
torch==2.1.2
accelerate==0.25.0

# Data processing
pandas==2.1.4
numpy==1.24.3
python-dotenv==1.0.0

# Web framework
fastapi==0.104.1
uvicorn==0.24.0
jinja2==3.1.2

# Utilities
pydantic==2.5.0
tiktoken==0.5.2
requests==2.31.0
beautifulsoup4==4.12.2
markdown==3.5.1

